requirements.txt
black==25.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
idna==3.11
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.1
python-dotenv==1.0.1
pytokens==0.3.0
requests==2.32.5
urllib3==2.6.2
